Animobj improvements
Now that stuff is working again (pay attention to project settings!)
took the time to improve file parsing for animation definitions. Changed
the format so that there will be less errors, such as eliminating
numframes since that can be calculated. Also added a termination
character to each sequence to simplify parsing.