Incorporated collision into entity
Now the entity class (actually player) has collision as a member, so
that just a call to update the player will also update the collision
database (which is external) at the same time.
The next step is to implement the actual collision routine using the
database...then fashion some sort of feedback to the affected entities
for response.